refactor: move SidebarNavItemProps to shared-components types
