+++ docs/docs/auto-docs/types/shared-components/SidebarNavItem/interface/interfaces/ISidebarNavItemProps.md
@@ -1,0 +1,85 @@
+[Admin Docs](/)
+
+***
+
+# Interface: ISidebarNavItemProps
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L4)
+
+Interface for SidebarNavItem component props.
+
+## Properties
+
+### dataCy?
+
+> `optional` **dataCy**: `string`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:13](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L13)
+
+***
+
+### hideDrawer
+
+> **hideDrawer**: `boolean`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L9)
+
+***
+
+### icon
+
+> **icon**: `ReactNode`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L6)
+
+***
+
+### iconType?
+
+> `optional` **iconType**: `"svg"` \| `"react-icon"`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L12)
+
+***
+
+### label
+
+> **label**: `string`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:7](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L7)
+
+***
+
+### onClick()?
+
+> `optional` **onClick**: () => `void`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L10)
+
+#### Returns
+
+`void`
+
+***
+
+### testId
+
+> **testId**: `string`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L8)
+
+***
+
+### to
+
+> **to**: `string`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:5](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L5)
+
+***
+
+### useSimpleButton?
+
+> `optional` **useSimpleButton**: `boolean`
+
+Defined in: [src/types/shared-components/SidebarNavItem/interface.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarNavItem/interface.ts#L11)
